--- a/tox.ini
+++ b/tox.ini
@@ -7,4 +7,4 @@ commands = py.test -v --capture=sys
 
 [flake8]
 ignore = E501,W503
-exclude = venv,build,.tox
+exclude = venv,build,.tox,setup.py,fastentrypoints.py