Before:
```bash
thefuck git br THEFUCK_ARGUMENT_PLACEHOLDER -y  0,36s user 0,02s system 97% cpu 0,383 total
thefuck --alias  0,29s user 0,01s system 99% cpu 0,303 total
```
After:
```bash
thefuck git br THEFUCK_ARGUMENT_PLACEHOLDER -y  0,19s user 0,03s system 94% cpu 0,234 total
thefuck --alias  0,13s user 0,02s system 99% cpu 0,149 total
```